Generate RIF-Seq test data from a cDNA FASTA
Use test data generated from a cDNA FASTA instead of modified real data,
in order to supply a small reference for testing purposes and still
yield results without errors. Update documentation to reflect new
changes, as well as with a detailed description on how the data and
reference was generated.