Store the LAST index in a single archive file.
I also moved the location of the data to the folder
`data/genomics/sarscov2/genome/alignment/last/` so that I can reuse it
path for further test data, in particular the trained parameter file
generated by the `last/train` module that will be submitted once the
test index is available.

I assume that this branch will be rebased when it will be merged.