Example sequence alignment in MAF format using sarscov2 test data.
The SARS-CoV-2 genome MT192765.1 was aligned to the `contigs.fasta`
example set of contig sequences provided in this repository, using the
LAST aligner.  The resulting file, in Multiple Alignment File (MAF)
format, is needed for the test suite of submodules of `last`, such as
`last/mafswap`, `last/mafconvert`, or `last/split`, that we are
preparing to submit to the nf-core/modules repository (see Issue #464).

While the file is small, I provide it compressed with gzip as typical
MAF files can be large, therefore modules and their tests should be
written to received them compressed.