Example sequence alignment in tabular format using sarscov2 test data.
The SARS-CoV-2 genome MT192765.1 was aligned to the `contigs.fasta`
example set of contig sequences provided in this repository, using the
LAST aligner.  The resulting file in Multiple Alignment File (MAF),
`contigs.genome.maf.gz` was converted to tabular format by the
`last/mafconvert` submodule.  This file will be needed to test
the `last/dotplot` submodule for graphical display of pairwise
alignments. (See nf-core/modules#464 for a tracking of the submission
of all `last/` submodules).